Fug bixes xD
Fixed a bug in the /fetch command.  It was using the wrong steam id.  It was using the id of the person who issued the command instead of the person being teleported.  This would occasionally result in the teleportee being temp banned.  oops!

Re-arranged when custom code is loaded and tweaked the reload code so that a fauly in the custom code does not cause the bot to fail to load.  However since the custom code has hooks everywhere it will cause most bot commands to not work if there is a syntax error in gmsg_custom.lua.  Commands such as /reload code are not effected since the hook is now further down the code.  The bot will report the status of the code reload and the custom code reload separately and reports which file contains the fault.  It is up to you to fix your custom code changes to get your bot working right.

Increased the size of some fields in the database that hold strings such as names for stuff as some of you like really really long names.

Added deleteLine()s in many places in match_all.lua to stop the bot outputting most of the telnet lines to Mudlet's console window.  This can be re-enabled by putting the bot into debug mode, which you currently do by editing a single debug line near the top.

+++ mudlet/profiles/newbot/scripts/timers/APITimer.lua
@@ -8,7 +8,7 @@
 --]]
 
 function APITimer()
-	local row, cursor, errorString, URL, outputFile
+	local row, cursor, errorString, url, outputFile, cmd
 
 	if not botman.dbConnected then
 		return
@@ -20,7 +20,7 @@ function APITimer()
 		row = cursor:fetch({}, "a")
 
 		if row then
-			URL = row.URL
+			url = row.URL
 			outputFile = row.OutputFile
 			conn:execute("delete from APIQueue where id = " .. row.id)
 
@@ -36,11 +36,16 @@ function APITimer()
 			metrics.commands = metrics.commands + 1
 
 			if server.logBotCommands then
-				logBotCommand(botman.serverTime, URL)
+				logBotCommand(botman.serverTime, url)
 			end
 
-			os.remove(outputFile)
-			downloadFile(outputFile, URL)
+			if not string.find(url, "#") then
+				os.remove(outputFile)
+				downloadFile(outputFile, url)
+			else
+				cmd = string.sub(url, string.find(url, "command=") + 8, string.find(url, "&adminuser") - 1)
+				send(cmd)
+			end
 		end
 	end
 end
